requirements.txt
python-socketio[client]
pymongo~=4.10.1